create diffdrivecomponentbase
